Cast conc arrays to float32 before writing to tarball

+++ src/mitgcm_inputs/rbcs/__init__.py
@@ -182,7 +182,9 @@ def main(args: argparse.Namespace) -> int:
         for data_var in conc_and_relax.data_vars:
             if not data_var.startswith("conc"):
                 continue
-            current_conc_data = conc_and_relax[data_var].values.tobytes()
+            current_conc_data = (
+                conc_and_relax[data_var].values.astype("f4", copy=False).tobytes()
+            )
             current_conc = tarfile.TarInfo(f"conc/{data_var}.bin")
             set_tar_file_ownerships(current_conc)
             current_conc.mode = 0o644
